+++ NEWS.md
@@ -15,7 +15,7 @@
 * Added new multi-seasonal harmonic method with annual and semiannual components and optional timetrend
 * Fixed outdated dontrun examples
 * Added handling for ages and age-group bounds greater than or equal to 115 to avoid unplausible ages in linelist
-* Added option to exclude outbreak-associated cases from training and testing for GLM-based algorithms
+* Added option to exclude outbreak-associated cases from baseline fitting for GLM-based algorithms
 
 # SignalDetectionTool 0.9.0